+++ lib/query-builder-group.js
@@ -2,7 +2,7 @@
 
 const QueryBuilderError = require('./query-builder-error');
 const QueryBuilderFields = require('./query-builder-fields');
-const Utils = require('../utils');
+const Utils = require('./utils');
 
 class QueryBuilderGroup {
