+++ package.json
@@ -34,7 +34,6 @@
     "husky": "^2.4.1"
   },
   "files": [
-    "lib/",
-    "utils"
+    "lib/"
   ]
 }
